Add similar project slack-block-builder to README
Added [slack-block-builder](https://github.com/raycharius/slack-block-builder) to similar projects. Also added JSX-Slack to my README, in similar section.

Oh, and I've opened a pull request for BotKit, that adds JSX-Slack and Block Builder to the README as alternatives to the raw JSON

Thanks for your package! Seeing it was great motivation for me to finally open-source the library I've been using for my apps!